Image-Generator/requirements.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cycler==0.10.0
Flask==1.1.1
get==2019.4.13
gunicorn==20.0.4
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
kiwisolver==1.1.0
MarkupSafe==1.1.1
matplotlib==3.1.2
numpy==1.18.1
opencv-python==4.1.2.30
Pillow==7.0.0
post==2019.4.13
public==2019.4.13
pyparsing==2.4.6
python-dateutil==2.8.1
query-string==2019.4.13
request==2019.4.13
requests==2.22.0
scipy==1.4.1
six==1.14.0
soupsieve==1.9.5
urllib3==1.25.7
Werkzeug==0.16.0
wincertstore==0.2
